Consolidate React codebase
