--- app/src/App.css
@@ -1,5 +1,0 @@
-main {
-  padding: 1rem;
-  max-width: 800px;
-  margin: 0 auto;
-}
