--- app/src/App.jsx
@@ -1,27 +1,0 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import Navbar from './Navbar'
-import Home from './pages/Home'
-import Experience from './pages/Experience'
-import Education from './pages/Education'
-import Skills from './pages/Skills'
-import Projects from './pages/Projects'
-import Contact from './pages/Contact'
-import './App.css'
-
-export default function App() {
-  return (
-    <BrowserRouter>
-      <Navbar />
-      <main>
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/experience" element={<Experience />} />
-          <Route path="/education" element={<Education />} />
-          <Route path="/skills" element={<Skills />} />
-          <Route path="/projects" element={<Projects />} />
-          <Route path="/contact" element={<Contact />} />
-        </Routes>
-      </main>
-    </BrowserRouter>
-  )
-}
