--- app/src/Navbar.css
@@ -1,25 +1,0 @@
-.navbar {
-  position: sticky;
-  top: 0;
-  width: 100%;
-  padding: 1rem;
-  background: #fff;
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-  transition: box-shadow 0.3s ease;
-  z-index: 1000;
-}
-
-.navbar.scrolled {
-  box-shadow: 0 2px 4px rgba(0,0,0,0.2);
-}
-
-.navbar a {
-  color: #333;
-  text-decoration: none;
-}
-
-.navbar a.active {
-  font-weight: bold;
-}
