--- app/src/Navbar.jsx
@@ -1,28 +1,0 @@
-import { NavLink } from 'react-router-dom'
-import { useEffect, useState } from 'react'
-import './Navbar.css'
-
-export default function Navbar() {
-  const [scrolled, setScrolled] = useState(false)
-
-  useEffect(() => {
-    const onScroll = () => {
-      setScrolled(window.scrollY > 50)
-    }
-    window.addEventListener('scroll', onScroll)
-    return () => window.removeEventListener('scroll', onScroll)
-  }, [])
-
-  return (
-    <header className={`navbar${scrolled ? ' scrolled' : ''}`}> 
-      <nav>
-        <NavLink to="/" end>Home</NavLink>
-        <NavLink to="/experience">Experience</NavLink>
-        <NavLink to="/education">Education</NavLink>
-        <NavLink to="/skills">Skills</NavLink>
-        <NavLink to="/projects">Projects</NavLink>
-        <NavLink to="/contact">Contact</NavLink>
-      </nav>
-    </header>
-  )
-}
