--- app/src/pages/Skills.jsx
@@ -1,10 +1,0 @@
-import { motion as Motion } from 'framer-motion'
-
-export default function Skills() {
-  return (
-    <Motion.div initial={{ opacity: 0 }} animate={{ opacity: 1 }} exit={{ opacity: 0 }}>
-      <h2>Skills</h2>
-      <p>This is the Skills page.</p>
-    </Motion.div>
-  )
-}
