+++ my-portfolio/package.json
@@ -17,7 +17,10 @@
     "framer-motion": "^12.23.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-type-animation": "^3.2.0"
+    "react-type-animation": "^3.2.0",
+    "@react-three/fiber": "^9.2.0",
+    "three": "^0.178.0",
+    "react-glitch-effect": "^1.0.8"
   },
   "devDependencies": {
     "@eslint/js": "^9.29.0",
