+++ my-portfolio/src/App.tsx
@@ -1,35 +1,30 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import Box from '@mui/material/Box';
+import { ThemeProvider } from '@mui/material/styles';
+import theme from './theme';
+import Starfield from './components/Starfield';
+import Intro from './components/Intro';
+import AboutMe from './components/AboutMe';
+import Skills from './components/Skills';
+import Projects from './components/Projects';
+import Certifications from './components/Certifications';
+import NerdSection from './components/NerdSection';
+import InteractiveCV from './components/InteractiveCV';
+import Contact from './components/Contact';
 
-function App() {
-  const [count, setCount] = useState(0)
-
+export default function App() {
   return (
-    <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
-    </>
-  )
+    <ThemeProvider theme={theme}>
+      <Starfield />
+      <Box sx={{ p: 2 }}>
+        <Intro />
+        <AboutMe />
+        <Skills />
+        <Projects />
+        <Certifications />
+        <NerdSection />
+        <InteractiveCV />
+        <Contact />
+      </Box>
+    </ThemeProvider>
+  );
 }
-
-export default App
