+++ my-portfolio/src/components/NerdSection.tsx
