+++ my-portfolio/src/index.css
@@ -1,5 +1,5 @@
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  font-family: 'Fira Code', monospace;
   line-height: 1.5;
   font-weight: 400;
