--- src/App.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import Starfield from './components/Starfield';
-
-const App: React.FC = () => {
-  return (
-    <>
-      <Starfield />
-      <div id="content">
-        {/* Add your site content here */}
-      </div>
-    </>
-  );
-};
-
-export default App;
